Extract pipelines api module
